Implement automatic dark mode, matching the system color scheme.
Using:

- The `prefers-color-scheme` media feature, widely available since Jan. 2020: https://developer.mozilla.org/en-US/docs/Web/CSS/Reference/At-rules/@media/prefers-color-scheme
- The `color-scheme` property, widely available since Jan. 2022: https://developer.mozilla.org/en-US/docs/Web/CSS/Reference/Properties/color-scheme
- `color-mix(…)`, widely supported since May 2023: https://developer.mozilla.org/en-US/docs/Web/CSS/Reference/Values/color_value/color-mix

In light mode, the appearance remains nearly as it was before.

A toggle is specifically *not* implemented, as that requires significant design aand implemenation work for good UX. Additionally, persistent state is less useful than it is for fully-featured PWAs, since 1) devs using `permission.site` will often need to do so across many devices/browsers/browser profile, and 2) testing is more liable to reset site state.

There are also several opportunities to clean up and/or modernize the CSS, but this is left out of scope here.